Process::Create() not searching PATH for partial executable name
Date: 	Wed, 27 Dec 2006 03:55:22 -0800
To: 	bug-libwin32@rt.cpan.org
From: 	Greg Ercolano <erco@seriss.com>

OK, after reading perlguts and perlxs (woof) I think I get it.

The right approach might be to allow 'undef' on the perl side
to be used where NULL would be needed on the C side to pass to
CreateProcess().

So it seems the simple solution might be to make this simple two line
change to Process.xs:

```
--- Process/Process.xs  2005-09-17 12:36:38.000000000 -0700
+++ Process/Process.xs.erco     2006-12-27 03:19:29.000000000 -0800
@@ -283,6 +283,8 @@
      DWORD flags
      char *curdir
  CODE:
+    appname = SvOK(ST(1)) ? appname : 0;
+    cmdline = SvOK(ST(2)) ? cmdline : 0;
      RETVAL = Create(cP, appname, cmdline, inherit, flags, curdir);
  OUTPUT:
      cP
```

..which then makes this work correctly:

    Process::Create($pobj, undef, "netstat -an", 1, NORMAL_PRIORITY_CLASS, ".");

The docs should cover this, so I've included some recommended text